Added referrer stats

+++ app/models/domain_log_group.rb
@@ -5,7 +5,7 @@ class DomainLogGroup < DomainModel
   named_scope :for_target, lambda { |type, id| {:conditions => {:target_type => type, :target_id => id}} }
   named_scope :with_type, lambda { |type| {:conditions => {:stat_type => type}} }
   named_scope :started_at, lambda { |time| {:conditions => {:started_at => time}} }
-  named_scope :with_duration, lambda { |duration| {:conditions => {:duration => duration}} }
+  named_scope :with_duration, lambda { |duration| {:conditions => {:duration => duration.to_i}} }
 
   def expired?
     self.expires_at && self.expires_at < Time.now
@@ -14,29 +14,46 @@ class DomainLogGroup < DomainModel
   def self.stats(target, from, duration, intervals, opts={})
     groups = []
     (1..intervals).each do |i|
-      scope = yield from, duration
-      groups << self.fetch_group(target, from, duration, scope, opts)
+      group = self.find_group(target, from, duration, opts)
+      if group
+        groups << group
+      else
+        scope = yield from, duration
+        groups << self.create_group(target, from, duration, scope, opts)
+      end
       from += duration
     end
     groups
   end
 
-  def self.fetch_group(target, started_at, duration, scope, opts={})
+  def self.find_group(target, started_at, duration, opts={})
     target_type = target.is_a?(String) ? target : target.class.to_s
     target_id = target.is_a?(String) ? nil : target.id
-    duration = duration.to_i
 
-    # NOTE: remember to optimize, need to check for the stats using target_id nil first, then if not results check for target_id if set
-    group_scope = DomainLogGroup.started_at(started_at).with_duration(duration).for_target(target_type, target_id)
-    group_scope = group_scope.with_type(opts[:type]) if opts[:type]
-    group = group_scope.first
+    group = DomainLogGroup.started_at(started_at).with_duration(duration).for_target(target_type, nil).with_type(opts[:type]).first
 
     if group && group.expired?
-      group.destroy
-      group = nil
+      if target_id.nil?
+        group.destroy
+        return nil
+      end
     end
 
-    return group if group
+    if group.nil? && target_id
+      group = DomainLogGroup.started_at(started_at).with_duration(duration).for_target(target_type, target_id).with_type(opts[:type]).first
+      if group && group.expired?
+        group.destroy
+        return nil
+      end
+    end
+
+    group.target_id = target_id if group && target_id
+    group
+  end
+
+  def self.create_group(target, started_at, duration, scope, opts={})
+    target_type = target.is_a?(String) ? target : target.class.to_s
+    target_id = target.is_a?(String) ? nil : target.id
 
     results = scope.find :all
 
@@ -45,7 +62,7 @@ class DomainLogGroup < DomainModel
       :target_id => target_id,
       :stat_type => opts[:type],
       :started_at => started_at,
-      :duration => duration
+      :duration => duration.to_i
     }
 
     attributes[:expires_at] = 5.minutes.since if (started_at + duration) > Time.now
@@ -56,5 +73,10 @@ class DomainLogGroup < DomainModel
     end
 
     group
+  end
+
+  def target_stats(target_id=nil)
+    target_id ||= self.target_id
+    self.domain_log_stats.find(:all, :conditions => {:target_id => target_id})
   end
 end
